% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afficher_resume_verifications.R
\name{afficher_resume_verifications}
\alias{afficher_resume_verifications}
\title{Afficher un resume des verifications de coherence}
\usage{
afficher_resume_verifications(verifications)
}
\arguments{
\item{verifications}{Liste des verifications retournee par verifier_coherence_donnees}
}
\description{
Afficher un resume des verifications de coherence
}
