% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_labels_base_menage.R
\name{ajouter_labels_base_menage}
\alias{ajouter_labels_base_menage}
\title{Ajouter les labels aux variables de la base menage}
\usage{
ajouter_labels_base_menage(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec labels ajoutes
}
\description{
Ajouter les labels aux variables de la base menage
}
