% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_labels_modalites.R
\name{ajouter_labels_modalites}
\alias{ajouter_labels_modalites}
\title{Ajouter les labels de modalites aux variables categoriques}
\usage{
ajouter_labels_modalites(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales}
}
\value{
Dataframe avec modalites labellisees
}
\description{
Ajouter les labels de modalites aux variables categoriques
}
