% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appliquer_conversion_poids.R
\name{appliquer_conversion_poids}
\alias{appliquer_conversion_poids}
\title{Appliquer la table de conversion pour obtenir les poids en kg}
\usage{
appliquer_conversion_poids(cereales, TableConversionP2)
}
\arguments{
\item{cereales}{Dataframe des cereales}

\item{TableConversionP2}{Table de conversion}
}
\value{
Dataframe avec les quantites en kg
}
\description{
Appliquer la table de conversion pour obtenir les poids en kg
}
