% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_calories.R
\name{calculer_calories}
\alias{calculer_calories}
\title{Calculer les calories consommees}
\usage{
calculer_calories(data, conversion_cal)
}
\arguments{
\item{data}{Dataframe}

\item{conversion_cal}{Table de conversion caloriques}
}
\value{
Dataframe avec les calories calculees
}
\description{
Calculer les calories consommees
}
