% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_calories_par_tete.R
\name{calculer_calories_par_tete}
\alias{calculer_calories_par_tete}
\title{Calculer les calories par tete}
\usage{
calculer_calories_par_tete(data, membres_menage)
}
\arguments{
\item{data}{Dataframe avec les calories individuelles}

\item{membres_menage}{Table des membres du menage}
}
\value{
Dataframe avec les calories par tete calculees
}
\description{
Calculer les calories par tete
}
