% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_depenses_consommation.R
\name{calculer_depenses_consommation}
\alias{calculer_depenses_consommation}
\title{Calculer les depenses de consommation}
\usage{
calculer_depenses_consommation(cereales, baseVU)
}
\arguments{
\item{cereales}{Dataframe des cereales}

\item{baseVU}{Base des valeurs unitaires de reference}
}
\value{
Dataframe avec les valeurs de consommation calculees
}
\description{
Calculer les depenses de consommation
}
