% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_depenses_par_tete.R
\name{calculer_depenses_par_tete}
\alias{calculer_depenses_par_tete}
\title{Calculer les depenses de consommation par tete}
\usage{
calculer_depenses_par_tete(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec depenses par tete calculees
}
\description{
Calculer les depenses de consommation par tete
}
