% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_frequence_achat.R
\name{calculer_frequence_achat}
\alias{calculer_frequence_achat}
\title{Calculer la frequence d'achat}
\usage{
calculer_frequence_achat(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales}
}
\value{
Dataframe avec la frequence d'achat calculee
}
\description{
Calculer la frequence d'achat
}
