% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_indice_diversite_simpson.R
\name{calculer_indice_diversite_simpson}
\alias{calculer_indice_diversite_simpson}
\title{Calculer l'indice de diversite cerealiere de Simpson}
\usage{
calculer_indice_diversite_simpson(data)
}
\arguments{
\item{data}{Dataframe au niveau produit-menage}
}
\value{
Dataframe avec indices de diversite par menage
}
\description{
Calculer l'indice de diversite cerealiere de Simpson
}
