% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_rendement_kcal_fcfa.R
\name{calculer_rendement_kcal_fcfa}
\alias{calculer_rendement_kcal_fcfa}
\title{Calculer le rendement kcal par FCFA depense}
\usage{
calculer_rendement_kcal_fcfa(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec rendement calcule
}
\description{
Calculer le rendement kcal par FCFA depense
}
