% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_taux_autoconsommation.R
\name{calculer_taux_autoconsommation}
\alias{calculer_taux_autoconsommation}
\title{Calculer le taux d'autoconsommation}
\usage{
calculer_taux_autoconsommation(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec taux d'autoconsommation calcule
}
\description{
Calculer le taux d'autoconsommation
}
