% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_valeurs_unitaires.R
\name{calculer_valeurs_unitaires}
\alias{calculer_valeurs_unitaires}
\title{Calculer les valeurs unitaires}
\usage{
calculer_valeurs_unitaires(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales}
}
\value{
Dataframe avec la variable vu ajoutee
}
\description{
Calculer les valeurs unitaires
}
