% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographie_ehcvm.R
\name{cartographie_ehcvm}
\alias{cartographie_ehcvm}
\title{Fonction complete de cartographie pour les donnees EHCVM du Senegal}
\usage{
cartographie_ehcvm(
  chemin_shapefile,
  donnees_menage,
  chemin_sortie,
  niveau_admin = "region",
  var_depense = "Depense_de_consommation_par_tete",
  var_calories = "kcal_par_tete",
  var_autoconsommation = "taux_autoconsommation",
  var_diversite = "indice_div_cereales"
)
}
\arguments{
\item{chemin_shapefile}{Chemin vers le fichier shapefile (.shp)}

\item{donnees_menage}{Dataframe Base_menage}

\item{chemin_sortie}{Dossier de sortie pour les resultats}

\item{niveau_admin}{Niveau administratif ("region" ou "departement")}

\item{var_depense}{Nom de la variable des depenses}

\item{var_calories}{Nom de la variable calorique}

\item{var_autoconsommation}{Nom de la variable d'autoconsommation}

\item{var_diversite}{Nom de la variable de diversite}
}
\value{
Liste contenant tous les resultats et cartes
}
\description{
Cette fonction integre toutes les etapes de cartographie :
1. Chargement des donnees geographiques
2. Preparation des donnees statistiques
3. Creation des cartes choroplethes
4. Analyse spatiale
5. Export des resultats
}
