% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographier_diversite_cereales.R
\name{cartographier_diversite_cereales}
\alias{cartographier_diversite_cereales}
\title{Carte de l'indice moyen de diversite des cereales par region}
\usage{
cartographier_diversite_cereales(base_menage, regions_sf)
}
\arguments{
\item{base_menage}{Un data.frame contenant au moins les colonnes `region` et `indice_div_cereales`.}

\item{shapefile_path}{Chemin d'acces complet au fichier shapefile des regions (.shp).}
}
\value{
Une carte ggplot2 affichant l'indice moyen par region.
}
\description{
Cette fonction calcule l'indice moyen de diversite des cereales par region
a partir de la base menage, puis le projette sur une carte a partir d'un fichier shapefile des regions.
}
