% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoriser_niveau_consommation.R
\name{categoriser_niveau_consommation}
\alias{categoriser_niveau_consommation}
\title{Categoriser le niveau de consommation en cereales}
\usage{
categoriser_niveau_consommation(data, seuil_min = 1114, cv_seuil = 0.29)
}
\arguments{
\item{data}{Dataframe avec kcal_par_tete}

\item{seuil_min}{Seuil minimum (defaut: 1114)}

\item{cv_seuil}{Coefficient de variation (defaut: 0.29)}
}
\value{
Dataframe avec la categorisation
}
\description{
Categoriser le niveau de consommation en cereales
}
