% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charger_donnees_ehcvm.R
\name{charger_donnees_ehcvm}
\alias{charger_donnees_ehcvm}
\title{Charger et preparer les donnees EHCVM}
\usage{
charger_donnees_ehcvm(chemin_donnees)
}
\arguments{
\item{chemin_donnees}{Chemin vers le dossier contenant les fichiers de donnees}
}
\value{
Liste contenant toutes les bases chargees
}
\description{
Charger et preparer les donnees EHCVM
}
