% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corriger_outliers_winsorisation.R
\name{corriger_outliers_winsorisation}
\alias{corriger_outliers_winsorisation}
\title{Corriger les outliers par winsorisation}
\usage{
corriger_outliers_winsorisation(data, var_name)
}
\arguments{
\item{data}{Dataframe}

\item{var_name}{Nom de la variable a corriger}
}
\value{
Dataframe avec les outliers corriges
}
\description{
Corriger les outliers par winsorisation
}
