% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_base_menage.R
\name{creer_base_menage}
\alias{creer_base_menage}
\title{Creer la base menage a partir des donnees individuelles}
\usage{
creer_base_menage(data)
}
\arguments{
\item{data}{Dataframe avec donnees au niveau produit-menage}
}
\value{
Dataframe agrege au niveau menage
}
\description{
Creer la base menage a partir des donnees individuelles
}
