% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_base_vu.R
\name{creer_base_vu}
\alias{creer_base_vu}
\title{Creer la base des valeurs unitaires de reference}
\usage{
creer_base_vu(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales avec VU calculees}
}
\value{
Dataframe avec les VU de reference par produit-unite-taille
}
\description{
Creer la base des valeurs unitaires de reference
}
