% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_tableau_bord.R
\name{creer_tableau_bord}
\alias{creer_tableau_bord}
\title{Creer un tableau de bord interactif simple}
\usage{
creer_tableau_bord(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Liste de graphiques pour tableau de bord
}
\description{
Creer un tableau de bord interactif simple
}
