% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_typologie_taille_menage.R
\name{creer_typologie_taille_menage}
\alias{creer_typologie_taille_menage}
\title{Creer une typologie de la taille du menage}
\usage{
creer_typologie_taille_menage(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec typologie de taille ajoutee
}
\description{
Creer une typologie de la taille du menage
}
