% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definir_source_dominante.R
\name{definir_source_dominante}
\alias{definir_source_dominante}
\title{Definir la source dominante de consommation}
\usage{
definir_source_dominante(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Dataframe avec source dominante definie
}
\description{
Definir la source dominante de consommation
}
