% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_outliers_iqr.R
\name{detecter_outliers_iqr}
\alias{detecter_outliers_iqr}
\title{Detecter les outliers par la methode IQR}
\usage{
detecter_outliers_iqr(data, var_name)
}
\arguments{
\item{data}{Dataframe}

\item{var_name}{Nom de la variable (caractere)}
}
\value{
Indices des outliers
}
\description{
Detecter les outliers par la methode IQR
}
