% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostiquer_donnees.R
\name{diagnostiquer_donnees}
\alias{diagnostiquer_donnees}
\title{Diagnostiquer les problemes potentiels dans les donnees}
\usage{
diagnostiquer_donnees(donnees)
}
\arguments{
\item{donnees}{Liste des donnees brutes}
}
\value{
Liste des diagnostics
}
\description{
Diagnostiquer les problemes potentiels dans les donnees
}
