% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichir_donnees_geo.R
\name{enrichir_donnees_geo}
\alias{enrichir_donnees_geo}
\title{Enrichir les donnees avec les informations geographiques}
\usage{
enrichir_donnees_geo(data, Ehcvm_all)
}
\arguments{
\item{data}{Dataframe principal}

\item{Ehcvm_all}{Base complete EHCVM}
}
\value{
Dataframe enrichi avec les informations geographiques
}
\description{
Enrichir les donnees avec les informations geographiques
}
