% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemple_utilisation_complete.R
\name{exemple_utilisation_complete}
\alias{exemple_utilisation_complete}
\title{Exemple d'utilisation complete du package}
\usage{
exemple_utilisation_complete(chemin_donnees, chemin_sortie)
}
\arguments{
\item{chemin_donnees}{Chemin vers le dossier contenant les fichiers de donnees}

\item{chemin_sortie}{Chemin vers le dossier de sortie des resultats}
}
\description{
Cette fonction montre comment utiliser l'ensemble des fonctions
pour une analyse complete des donnees EHCVM
}
\examples{
\dontrun{
# Utilisation du pipeline complet
resultats <- pipeline_analyse_ehcvm(
  chemin_donnees = "C:/Users/HP/Desktop/data",
  chemin_sortie = "C:/Users/HP/Desktop/outputs"
)

# Ou utilisation etape par etape
donnees <- charger_donnees_ehcvm("C:/Users/HP/Desktop/data")
cereales <- preparer_colonnes_cereales(donnees$cereales)
cereales <- ajouter_labels_modalites(cereales)
# ... suite des etapes
}
}
