% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter_vers_excel.R
\name{exporter_vers_excel}
\alias{exporter_vers_excel}
\title{Exporter les resultats vers Excel avec plusieurs onglets}
\usage{
exporter_vers_excel(resultats, chemin_excel)
}
\arguments{
\item{resultats}{Liste des resultats}

\item{chemin_excel}{Chemin du fichier Excel de sortie}
}
\description{
Exporter les resultats vers Excel avec plusieurs onglets
}
