% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport_synthese.R
\name{generer_rapport_synthese}
\alias{generer_rapport_synthese}
\title{Generer un rapport de synthese des analyses}
\usage{
generer_rapport_synthese(resultats, chemin_rapport = NULL)
}
\arguments{
\item{resultats}{Liste retournee par pipeline_analyse_ehcvm}

\item{chemin_rapport}{Chemin pour sauvegarder le rapport}
}
\description{
Generer un rapport de synthese des analyses
}
