% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer_valeurs_manquantes.R
\name{imputer_valeurs_manquantes}
\alias{imputer_valeurs_manquantes}
\title{Imputer les valeurs manquantes par k-NN}
\usage{
imputer_valeurs_manquantes(data, vars_a_imputer, vars_explicatives, k = 5)
}
\arguments{
\item{data}{Dataframe avec valeurs manquantes}

\item{vars_a_imputer}{Variables a imputer}

\item{vars_explicatives}{Variables explicatives pour l'imputation}

\item{k}{Nombre de voisins (defaut: 5)}
}
\value{
Dataframe avec valeurs imputees
}
\description{
Imputer les valeurs manquantes par k-NN
}
