% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_hsm_safe.R
\name{mode_hsm_safe}
\alias{mode_hsm_safe}
\title{Calculer le mode en utilisant la methode HSM}
\usage{
mode_hsm_safe(x)
}
\arguments{
\item{x}{Vecteur numerique}
}
\value{
Valeur du mode ou NA si impossible a calculer
}
\description{
Calculer le mode en utilisant la methode HSM
}
