% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_analyse_ehcvm.R
\name{pipeline_analyse_ehcvm}
\alias{pipeline_analyse_ehcvm}
\title{Pipeline complet d'analyse des donnees EHCVM}
\usage{
pipeline_analyse_ehcvm(chemin_donnees, chemin_sortie)
}
\arguments{
\item{chemin_donnees}{Chemin vers le dossier des donnees}

\item{chemin_sortie}{Chemin vers le dossier de sortie}
}
\value{
Liste contenant toutes les bases finales
}
\description{
Pipeline complet d'analyse des donnees EHCVM
}
