% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barres_groupees.R
\name{plot_barres_groupees}
\alias{plot_barres_groupees}
\title{Creer un graphique en barres groupees pour deux variables categorielles}
\usage{
plot_barres_groupees(data, var_x, var_fill, titre = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{var_x}{Variable pour l'axe x}

\item{var_fill}{Variable pour le remplissage}

\item{titre}{Titre du graphique}
}
\value{
Objet ggplot
}
\description{
Creer un graphique en barres groupees pour deux variables categorielles
}
