% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barres_stat_par_groupe.R
\name{plot_barres_stat_par_groupe}
\alias{plot_barres_stat_par_groupe}
\title{Creer un graphique en barres pour une variable numerique par groupe}
\usage{
plot_barres_stat_par_groupe(
  data,
  var_num,
  var_groupe,
  stat_fonction = "mean",
  titre = NULL
)
}
\arguments{
\item{data}{Dataframe}

\item{var_num}{Variable numerique}

\item{var_groupe}{Variable de groupement}

\item{stat_fonction}{Fonction statistique (mean, median, sum)}

\item{titre}{Titre du graphique}
}
\value{
Objet ggplot
}
\description{
Creer un graphique en barres pour une variable numerique par groupe
}
