% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot_variable.R
\name{plot_boxplot_variable}
\alias{plot_boxplot_variable}
\title{Creer un boxplot pour une variable}
\usage{
plot_boxplot_variable(data, var, titre = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{var}{Variable a representer}

\item{titre}{Titre du graphique (optionnel)}
}
\value{
Objet ggplot
}
\description{
Creer un boxplot pour une variable
}
