% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_camembert.R
\name{plot_camembert}
\alias{plot_camembert}
\title{Creer un graphique en camembert pour une variable categorique}
\usage{
plot_camembert(data, var, titre = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{var}{Variable categorique (nom sans guillemets)}

\item{titre}{Titre du graphique}
}
\value{
Objet ggplot
}
\description{
Creer un graphique en camembert pour une variable categorique
}
