% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogramme.R
\name{plot_histogramme}
\alias{plot_histogramme}
\title{Creer un histogramme pour une variable}
\usage{
plot_histogramme(data, var, titre = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{var}{Variable a representer}

\item{titre}{Titre du graphique (optionnel)}
}
\value{
Objet ggplot
}
\description{
Creer un histogramme pour une variable
}
