% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_colonnes_cereales.R
\name{preparer_colonnes_cereales}
\alias{preparer_colonnes_cereales}
\title{Renommer et labelliser les colonnes de la base cereales}
\usage{
preparer_colonnes_cereales(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales}
}
\value{
Dataframe avec colonnes renommees et labellisees
}
\description{
Renommer et labelliser les colonnes de la base cereales
}
