% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumer_variable.R
\name{resumer_variable}
\alias{resumer_variable}
\title{Resumer une variable numerique}
\usage{
resumer_variable(data, var)
}
\arguments{
\item{data}{Dataframe}

\item{var}{Variable a resumer (nom sans guillemets)}
}
\value{
Dataframe avec les statistiques descriptives
}
\description{
Resumer une variable numerique
}
