% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauvegarder_csv_utf8.R
\name{sauvegarder_csv_utf8}
\alias{sauvegarder_csv_utf8}
\title{Sauvegarder un dataframe en CSV avec encodage UTF-8}
\usage{
sauvegarder_csv_utf8(data, chemin_fichier)
}
\arguments{
\item{data}{Dataframe a sauvegarder}

\item{chemin_fichier}{Chemin complet du fichier de sortie}
}
\description{
Sauvegarder un dataframe en CSV avec encodage UTF-8
}
