% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_coherence_resultats.R
\name{tester_coherence_resultats}
\alias{tester_coherence_resultats}
\title{Executer des tests de coherence sur les resultats}
\usage{
tester_coherence_resultats(Base_menage)
}
\arguments{
\item{Base_menage}{Dataframe de la base menage}
}
\value{
Liste des resultats de tests
}
\description{
Executer des tests de coherence sur les resultats
}
