% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traiter_valeurs_manquantes_sources.R
\name{traiter_valeurs_manquantes_sources}
\alias{traiter_valeurs_manquantes_sources}
\title{Traiter les valeurs manquantes pour les variables de source}
\usage{
traiter_valeurs_manquantes_sources(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
Dataframe avec valeurs manquantes traitees
}
\description{
Traiter les valeurs manquantes pour les variables de source
}
