% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valider_structure_donnees.R
\name{valider_structure_donnees}
\alias{valider_structure_donnees}
\title{Valider la structure des donnees d'entree}
\usage{
valider_structure_donnees(donnees)
}
\arguments{
\item{donnees}{Liste des donnees chargees}
}
\value{
TRUE si la structure est valide, erreur sinon
}
\description{
Valider la structure des donnees d'entree
}
