% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifier_coherence_donnees.R
\name{verifier_coherence_donnees}
\alias{verifier_coherence_donnees}
\title{Verifier la coherence interne des donnees}
\usage{
verifier_coherence_donnees(cereales)
}
\arguments{
\item{cereales}{Dataframe des cereales}
}
\value{
Liste des verifications avec les lignes problematiques
}
\description{
Verifier la coherence interne des donnees
}
