% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifier_configuration.R
\name{verifier_configuration}
\alias{verifier_configuration}
\title{Verifier la configuration systeme requise}
\usage{
verifier_configuration()
}
\value{
TRUE si tout est OK, liste des problemes sinon
}
\description{
Verifier la configuration systeme requise
}
